client/requirements.txt
aiohttp>=3.9.0
pynacl>=1.5.0
aiosqlite>=0.19.0
qrcode>=7.4
Pillow>=10.0
